Added JavaDoc comments to Subtract, Multiply and Divide. Refactored to use floats and added more tests using floats.
git-svn-id: https://svn.cim.rhul.ac.uk/personal/zdac085/Calculator@39 c378c551-b4bd-4fd4-84fa-af6bf24c0965

+++ branches/RevPolishCalc/src/calc/Divide.java
@@ -1,18 +1,40 @@
 package calc;
 
+/**
+ * This class implements the operator interface, so the calculate function can return the answer of
+ * x / y.
+ *
+ * @author Marcus Messer
+ */
 public class Divide implements Operator {
 
+  /**
+   * This field is the current instance of the class.
+   */
   private static Divide instance = new Divide();
 
+  /**
+   * This method gets the current instance of the class.
+   * @return The current instance.
+   */
   public static Divide getInstance() {
     return instance;
   }
 
+  /**
+   * The method is an empty constructor, that can only be called from within this class.
+   */
   private Divide() {
   }
 
+  /**
+   * This method calculates x /y.
+   * @param x The first (left hand) number.
+   * @param y The second (right hand) number.
+   * @return The answer of x / y.
+   */
   @Override
-  public int calculate(int x, int y) {
+  public float calculate(float x, float y) {
     return x / y;
   }
 }
